e2e: parse sync_key_counts to decide on key sending
The server returns the count of one-time-keys that it has,
we parse that and detect if we need to send more.

It's a little tricky in that a server that doesn't have
any keys will send an empty response, so force sending in
that case.

If we get to the point of multiple types of keys this will
need reworking to make sure we only send the key types needed.

Signed-off-by: Dr. David Alan Gilbert <dave@treblig.org>